Code/requirements.txt
xmltodict==0.13.0
matplotlib==3.10.0
pyserial==3.5